Move `pkg/predicate` to `pkg/controllerutils/predicate`
