+++ pkg/scheduler/controller/shoot/scheduler.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	"github.com/gardener/gardener/pkg/predicate"
+	"github.com/gardener/gardener/pkg/controllerutils/predicate"
 	"github.com/gardener/gardener/pkg/scheduler/apis/config"
 
 	"sigs.k8s.io/controller-runtime/pkg/controller"
